refactor: use new file structure for Member API
